fix column shadow in spreadsheet for rendering rows
